+++ srcs/minishell.c
@@ -6,13 +6,13 @@
 /*   By: craimond <bomboclat@bidol.juis>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/04 17:09:22 by craimond          #+#    #+#             */
-/*   Updated: 2024/03/02 16:15:30 by craimond         ###   ########.fr       */
+/*   Updated: 2024/03/02 22:58:45 by craimond         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../headers/minishell.h"
 
-int	g_status;
+uint8_t	g_status;
 
 static void minishell_loop(void);
