+++ srcs/minishell_exit.c
@@ -6,7 +6,7 @@
 /*   By: craimond <bomboclat@bidol.juis>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/28 00:31:02 by craimond          #+#    #+#             */
-/*   Updated: 2024/03/02 19:06:51 by craimond         ###   ########.fr       */
+/*   Updated: 2024/03/02 22:57:55 by craimond         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -70,7 +70,7 @@ void	free_data(void)
 
 static void	close_all_fds(void)
 {
-	int		fd;
+	uint16_t	fd;
 
 	fd = 2; //senza chiudere stdin e stdout
 	while (++fd < MAX_FDS)
