+++ srcs/parser/parser.c
@@ -6,7 +6,7 @@
 /*   By: craimond <bomboclat@bidol.juis>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/06 17:58:27 by craimond          #+#    #+#             */
-/*   Updated: 2024/03/02 16:10:56 by craimond         ###   ########.fr       */
+/*   Updated: 2024/03/02 23:28:33 by craimond         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -62,7 +62,7 @@ static t_tree   *fill_tree(t_list *lexered_params, t_list *stop)
 
 static t_list	*skip_parenthesis(t_list *lexered_params)
 {
-	int32_t	    n_open;
+	uint16_t	n_open;
 	t_lexer		*elem;
 
 	n_open = 1;
@@ -80,7 +80,7 @@ static t_list	*skip_parenthesis(t_list *lexered_params)
 
 static t_list   *unskip_parenthesis(t_list *lexered_params)
 {
-    int32_t    n_close;
+    uint16_t   n_close;
     t_lexer    *elem;
 
     n_close = 1;
