+++ srcs/utils/general/ft_matrixsize.c
@@ -6,15 +6,15 @@
 /*   By: craimond <bomboclat@bidol.juis>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/05 10:41:04 by craimond          #+#    #+#             */
-/*   Updated: 2024/02/12 00:08:43 by craimond         ###   ########.fr       */
+/*   Updated: 2024/03/03 00:17:45 by craimond         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../../headers/minishell.h"
 
-size_t	ft_matrixsize(char **matrix)
+uint16_t	ft_matrixsize(char **matrix)
 {
-    size_t	size;
+    uint16_t	size;
 
     size = 0;
     if (!matrix)
